src/microservice-a/requirements.txt
Flask
dapr-dapr-clients
